OneDrive/Documentos/GSI/Super Paqui/whatsapp/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
numpy
python-multipart
gspread
google-auth
jinja2
httpx
python-dotenv
